+++ lib/object_inspector/inspectors_helper.rb
@@ -5,10 +5,20 @@
 # generating the inspection output.
 module ObjectInspector::InspectorsHelper
   # Calls {ObjectInspector::Inspector.inspect} on the passed in `object`,
-  # passing it the passed in `kwargs` (keyword arguments).
+  # passing through any keyword arguments.
   #
   # @return [String]
   def inspect(object = self, **)
     ObjectInspector::Inspector.inspect(object, **)
+  end
+
+  # Like {#inspect} but forces scope to `:all`. This (the bang (!) version) is
+  # considered the "more dangerous" version of {#inspect} in the sense that the
+  # `:all` scope may result in additional queries or extra processing--depending
+  # on how the inspect hooks are setup.
+  #
+  # @return [String]
+  def inspect!(object = self, **)
+    ObjectInspector::Inspector.inspect(object, **, scope: :all)
   end
 end